[MIG] hr_recruitment_phone: Migration to 18.0
