+++ checklog-odoo.cfg
@@ -1,3 +1,4 @@
 [checklog-odoo]
 ignore=
     WARNING.* 0 failed, 0 error\(s\).*
+    WARNING.* Attachment indexation of PDF documents is unavailable because the 'pdfminer' Python library cannot be found on the system
