+++ hr_recruitment_phone/__manifest__.py
@@ -4,7 +4,7 @@
 
 {
     "name": "HR Recruitment Phone",
-    "version": "16.0.1.0.0",
+    "version": "18.0.1.0.0",
     "category": "Phone",
     "license": "AGPL-3",
     "summary": "Validate phone numbers in HR Recruitment",
