+++ hr_recruitment_phone/models/__init__.py
@@ -1,1 +1,1 @@
-from . import hr_applicant
+from . import hr_candidate
