+++ hr_recruitment_phone/readme/CONTRIBUTORS.md
@@ -1,1 +1,2 @@
 - Alexis de Lattre \<<alexis.delattre@akretion.com>\>
+- Chau Le \<<chaulb@trobz.com>\>
