+++ hr_recruitment_phone/security/ir.model.access.csv
@@ -1,2 +1,2 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-callerid_hr_applicant_read,Read access on hr.applicant,hr_recruitment.model_hr_applicant,base_phone.group_callerid,1,0,0,0
+callerid_hr_candidate_read,Read access on hr.candidate,hr_recruitment.model_hr_candidate,base_phone.group_callerid,1,0,0,0
