+++ hr_recruitment_phone/tests/test_phone.py
@@ -2,19 +2,20 @@
 # @author: Alexis de Lattre <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from odoo.tests.common import TransactionCase
+from odoo.addons.base.tests.common import BaseCommon
 
 
-class TestRecruitmentPhone(TransactionCase):
-    def setUp(self):
-        super().setUp()
-        self.fr_country_id = self.env.ref("base.fr").id
-        self.phco = self.env["phone.common"]
-        self.env.company.write({"country_id": self.fr_country_id})
-        self.partner = self.env["res.partner"].create(
+class TestRecruitmentPhone(BaseCommon):
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls.fr_country_id = cls.env.ref("base.fr").id
+        cls.phco = cls.env["phone.common"]
+        cls.env.company.write({"country_id": cls.fr_country_id})
+        cls.partner = cls.env["res.partner"].create(
             {
                 "name": "Partner 0",
-                "country_id": self.fr_country_id,
+                "country_id": cls.fr_country_id,
                 "phone": "+33 4 78 32 32 32",
             }
         )
@@ -24,8 +25,6 @@ class TestRecruitmentPhone(TransactionCase):
         self.assertIsInstance(res, tuple)
         self.assertEqual(res[0], "res.partner")
         self.assertEqual(res[1], self.partner.id)
-        self.assertEqual(
-            res[2], self.partner.with_context(callerid=True).name_get()[0][1]
-        )
+        self.assertEqual(res[2], self.partner.with_context(callerid=True).display_name)
         res = self.phco.get_record_from_phone_number("0499889988")
         self.assertFalse(res)
